xdp refactor done
